feat: 算力点支持小数（Credits 改为 float64）
breaking change: User.Credits / ModelCost.Credits / CreditLog.Amount+Balance
从 int 改为 float64, 支持 0.5 等半价。

主要改动:
- model/{user,setting}.go: 5 个 int 字段改 float64
- service/{auth,settings}.go: 函数签名 + 返回值
- repository/user.go: Consume/RefundUserCredits 签名
- handler/{ai,auth}.go: readAIRequestCount + adjustUserCreditsRequest
- handler/ai.go: 请求体 N 字段 float64

数据库: SQLite 自动适配（整数列存浮点也没问题）, 不需要 schema 迁移。

测试:
- 12.7 算力设置成功
- 0.5 算力设置成功
- 流水 amount 正确（含负数）
- API 返回正确
- mini 模型 0.5 算力正常

兼容性:
- 前端 AdminUser.credits / AdminCreditLog.amount 等类型已经是 number
- 现有 int 数据兼容（无小数部分）
- 所有现有功能不变

+++ handler/ai.go
@@ -183,8 +183,8 @@ func readMultipartModel(body []byte, contentType string) string {
 	return ""
 }
 
-func readAIRequestCount(body []byte, contentType string) int {
-	count := 1
+func readAIRequestCount(body []byte, contentType string) float64 {
+	count := 1.0
 	if strings.HasPrefix(contentType, "multipart/form-data") {
 		_, params, err := mime.ParseMediaType(contentType)
 		if err != nil {
@@ -200,7 +200,7 @@ func readAIRequestCount(body []byte, contentType string) int {
 		}
 	} else {
 		var payload struct {
-			N int `json:"n"`
+			N float64 `json:"n"`
 		}
 		_ = json.Unmarshal(body, &payload)
 		count = payload.N
